SI-4558 refactor(logger): Отдельный файл для данных события для лога
