+++ OOP Patterns/aquarium/lib/commands/implementations/duty_commands.dart
@@ -1,5 +1,6 @@
 import 'package:aquarium/commands/command.dart';
 import 'package:aquarium/logger/base_logger.dart';
+import 'package:aquarium/logger/data/log_event_data.dart';
 import 'package:aquarium/logger/res/log_res.dart';
 import 'package:aquarium/pool/staff.dart';
