+++ OOP Patterns/aquarium/lib/fish/fish.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:aquarium/fish/strategy/react_pool_strategy.dart';
 import 'package:aquarium/logger/base_logger.dart';
+import 'package:aquarium/logger/data/log_event_data.dart';
 import 'package:aquarium/logger/res/log_res.dart';
 import 'package:aquarium/pool/observable.dart';
 import 'package:aquarium/pool/pool_state.dart';
