+++ OOP Patterns/aquarium/lib/logger/data/log_event_data.dart
@@ -1,0 +1,17 @@
+/// Сущность события для логирования
+class LogEventData {
+  /// Вреся наступления события
+  DateTime dateTime;
+
+  /// Описание события
+  String description;
+
+  /// Объект события
+  Object? object;
+
+  LogEventData({
+    required this.dateTime,
+    required this.description,
+    this.object,
+  });
+}
