+++ OOP Patterns/aquarium/lib/logger/description_stream_logger.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:aquarium/logger/base_logger.dart';
+import 'package:aquarium/logger/data/log_event_data.dart';
 import 'package:intl/intl.dart';
 
 /// Логгер который заносит описание событий в стрим
